renaming files with wrong case to support linux build
