+++ src/index.ts
@@ -30,7 +30,7 @@ import * as scrypt from './scrypt'
 import * as core from './core'
 import * as utils from './utils'
 import * as CONST from './consts'
-import { SDK } from './SDK/index'
+import { SDK } from './sdk/index'
 import { CompleterResult } from 'readline';
 
 class ONT {
