+++ src/smartcontract/data/idContract.avm
