+++ src/smartcontract/data/idContract_v0.2.avm
